+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog
 Version 4.0.0
 -------------
 
-- Fix ``neurom.load_morphologies`` to always pass resolved paths to Population.
+- Make ``neurom.core.Population`` resolve paths. Symlinks are resolved up until the
+  parent directory. If the file is symlinked, then it will not be resolved.
 - Mixed subtree processing can be used in morph_stats app via the use_subtrees flag.
 - ``neurom.view.[plot_tree|plot_tree3d|plot_soma|plot_soma3D]`` were hidden from the
   neurom.view module. They can still be imported from neurom.view.matplotlib_impl.